refactor(book-search): useSelectedSearchBook hook
