+++ app/constants.ts
@@ -1,0 +1,2 @@
+export const BOOK_PARAM = "book";
+export const QUERY_PARAM = "q";
